% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{adhocLogPhiSD}
\alias{adhocLogPhiSD}
\title{Helper function to calculate ad-hoc log(phi) replicate stdev
to be used for the ad-hoc filter.}
\usage{
adhocLogPhiSD(regionalCounts, params, condQuad, tp)
}
\arguments{
\item{regionalCounts}{the regional counts table}

\item{params}{the parameter sheet object}

\item{condQuad}{the set of select,nonselect,selWT and nonWT conditions}

\item{tp}{the current time point}
}
\value{
a vector of ad-hoc log(phi) sds.
}
\description{
Helper function to calculate ad-hoc log(phi) replicate stdev
to be used for the ad-hoc filter.
}
