% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{adjustTableDepth}
\alias{adjustTableDepth}
\title{Adjust dataframe 'covTable' with mean "effective" sequencing depth}
\usage{
adjustTableDepth(covTable, params)
}
\arguments{
\item{covTable}{dataframe with position-specific effective depth at each 
nucleotide position in each sample}

\item{params}{parameter sheet object}
}
\value{
table with adjusted mean effective depths
}
\description{
Adjust dataframe 'covTable' with mean "effective" sequencing depth
}
