% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{analyzeComplexity}
\alias{analyzeComplexity}
\title{Determine complexity of all varinats in allCounts; look at marginal frequency 
against how many unique contexts the variant is seen in}
\usage{
analyzeComplexity(allCounts, simplifiedMarginal)
}
\arguments{
\item{allCounts}{all counts table}

\item{simplifiedMarginal}{simplified marginal frequency table}
}
\value{
table with "complexity" vs marginal frequency
}
\description{
Determine complexity of all varinats in allCounts; look at marginal frequency 
against how many unique contexts the variant is seen in
}
