% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{biasCorrection}
\alias{biasCorrection}
\title{Run read-frequecy bias correction on logPhi}
\usage{
biasCorrection(msc)
}
\arguments{
\item{msc}{the enrichment table}
}
\value{
the bias-corrected enrichment (bce) and its stddev
}
\description{
Run read-frequecy bias correction on logPhi
}
