% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{bnl}
\alias{bnl}
\title{Baldi & Long's formula}
\usage{
bnl(pseudo.n, n, model.sd, empiric.sd)
}
\arguments{
\item{pseudo.n}{the number of pseudo-replicates (the weight of the prior)}

\item{n}{the number of empirical replicates}

\item{model.sd}{the prior expectation of the standard deviation}

\item{empiric.sd}{the empircal standard deviation}
}
\description{
Baldi & Long's formula
}
