% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildJointTable.R
\name{buildJointTable}
\alias{buildJointTable}
\title{run library quality control (QC)}
\usage{
buildJointTable(
  dataDir,
  paramFile = paste0(dataDir, "parameters.json"),
  mc.cores = 6,
  srOverride = FALSE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}

\item{mc.cores}{the number of CPU cores to use in parallel}

\item{srOverride}{the single-replicate override flag}
}
\value{
NULL. Results are written to file.
}
\description{
run library quality control (QC)
}
