% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{buildPlotMatrix}
\alias{buildPlotMatrix}
\title{Helper function to build layout for coverage map}
\usage{
buildPlotMatrix(ri, tpr, tileSets, lastTile)
}
\arguments{
\item{ri}{region index}

\item{tpr}{number of tiles per row}

\item{tileSets}{}

\item{lastTile}{number of the last tile in region}
}
\value{
table that indicates the grid positions of plots for region in order of drawing
}
\description{
Helper function to build layout for coverage map
}
