% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{buildSimplifiedMarginal}
\alias{buildSimplifiedMarginal}
\title{Build a simplified version of the marginal table with columns for variant, mean marginal  
frequency, marginal depth and tile, with 0-depth variants filtered out}
\usage{
buildSimplifiedMarginal(
  marginalSplitChanges,
  nsMarginalMeans,
  nsMarginalDepth,
  marginalTiles
)
}
\arguments{
\item{marginalSplitChanges}{table with split codon changes from marginal counts table}

\item{nsMarginalMeans}{table with mean marginal frequency across nonselect technical replicates}

\item{nsMarginalDepth}{table with mean marginal depth across nonselect technical replicates}

\item{marginalTiles}{table with tile assignments for each entry in marginal 
counts table}
}
\value{
table with variant, mean marginal frequency, marginal depth and tile
}
\description{
Build a simplified version of the marginal table with columns for variant, mean marginal  
frequency, marginal depth and tile, with 0-depth variants filtered out
}
