% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{calcDepthDrops}
\alias{calcDepthDrops}
\title{Calculate relative drops in effective sequencing depth for each variant}
\usage{
calcDepthDrops(xCounts, xTiles, depthTable, mc.cores = 6)
}
\arguments{
\item{xCounts}{marginal or all counts table}

\item{xTiles}{tile assignments for each variant in the table}

\item{depthTable}{raw depth table}

\item{mc.cores}{number of processes to run in parallel for multi-core processing}
}
\value{
a table listing the relative drop for each variant in each
condition-time-replicate combination
}
\description{
Calculate relative drops in effective sequencing depth for each variant
}
