% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{calcEnrichment}
\alias{calcEnrichment}
\title{calculate logPhi enrichment ratios}
\usage{
calcEnrichment(
  dataDir,
  inDir = NA,
  outDir = NA,
  paramFile = paste0(dataDir, "parameters.json"),
  mc.cores = 6,
  srOverride = FALSE,
  bnOverride = FALSE,
  bcOverride = FALSE,
  useWTfilter = FALSE,
  nbs = 10000,
  pessimistic = TRUE,
  useQuorum = FALSE,
  useLPDfilter = FALSE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{inDir}{input directory, defaults to subdirectory with latest timestamp ending in _mut_count.}

\item{outDir}{output directory, defaults to name of input directory with _QC tag attached.}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}

\item{mc.cores}{number of CPU cores to use in parallel}

\item{srOverride}{single replicate override}

\item{bnOverride}{bottleneck filter override}
}
\value{
NULL. Results are written to file.
}
\description{
performs error modelling and regularization, filtering and calculates logPhi
enrichment values
}
