% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calcPhi}
\alias{calcPhi}
\title{Calculate enrichment ratios (phi) and perform error propagation}
\usage{
calcPhi(msc)
}
\arguments{
\item{msc}{data.frame containing the output of the 'mean.sd.count()' function}
}
\value{
a data.frame containing phi, log(phi) and their respective stdevs
}
\description{
Calculate enrichment ratios (phi) and perform error propagation
}
