% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{calculateFits}
\alias{calculateFits}
\title{Calculate lambda and Poisson fits for each census}
\usage{
calculateFits(tile, tileCensi)
}
\arguments{
\item{tile}{tile number being analyzed}

\item{tileCensi}{table containing population breakdown of variant types within each tile}
}
\value{
vector of lambda and Poisson fits for census
}
\description{
Calculate lambda and Poisson fits for each census
}
