% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{charProfile}
\alias{charProfile}
\title{Get a contingency table of characters in a file}
\usage{
charProfile(filename)
}
\arguments{
\item{filename}{the file to analyze}
}
\value{
a contingency table of the characters in the file
}
\description{
Get a contingency table of characters in a file
}
