% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{checkCSV}
\alias{checkCSV}
\title{Checks if a file is a CSV file or throws errors}
\usage{
checkCSV(filename)
}
\arguments{
\item{filename}{the file to check}
}
\value{
nothing
}
\description{
Checks if a file is a CSV file or throws errors
}
