% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{codonAgreement}
\alias{codonAgreement}
\title{Draw plots to analyze agreement between equivalent codons}
\usage{
codonAgreement(scores, sCond, tp, params, outDir, srOverride)
}
\arguments{
\item{scores}{the score table}

\item{sCond}{currently active selection condition}

\item{tp}{currently active time point}

\item{params}{parameter sheet object}

\item{outdir}{output directory}
}
\value{
nothing. writes plots to output directory
}
\description{
Draw plots to analyze agreement between equivalent codons
}
