% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{countAttributes}
\alias{countAttributes}
\title{Output alignment quality plot with percentages of unmapped or mismapped reads
per sample}
\usage{
countAttributes(depthTable, inDir, outDir, pdftag)
}
\arguments{
\item{depthTable}{raw depth table}

\item{inDir}{input directory}

\item{outDir}{output directory}

\item{pdftag}{PDF tag string}

\item{params}{parameter sheet object}
}
\description{
Output alignment quality plot with percentages of unmapped or mismapped reads
per sample
}
