% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{csvParam2Json}
\alias{csvParam2Json}
\title{convert CSV input parameter file to JSON format}
\usage{
csvParam2Json(infile, outfile = NULL)
}
\arguments{
\item{infile}{the input CSV file}
}
\value{
nothing, but an output file will be written in the same directory as the input file
}
\description{
convert CSV input parameter file to JSON format
}
