% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{drawComplexity}
\alias{drawComplexity}
\title{Output complexity plot with the marginal frequency of variants plotted 
against the number of unique genetic contexts in which they were observed}
\usage{
drawComplexity(info, cplxPlot)
}
\arguments{
\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{cplxPlot}{table with "complexity" vs marginal frequency}
}
\description{
Output complexity plot with the marginal frequency of variants plotted 
against the number of unique genetic contexts in which they were observed
}
