% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{drawCoverageLegend}
\alias{drawCoverageLegend}
\title{Draw legend for coverage maps}
\usage{
drawCoverageLegend(wmThreshold, aaMarginal)
}
\arguments{
\item{wmThreshold}{well-measuredness threshold}

\item{aaMarginal}{marginal frequency table at amino acid-level}
}
\description{
Draw legend for coverage maps
}
