% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{drawCoverageMap}
\alias{drawCoverageMap}
\title{Output heatmap of marginal frequencies for all possible amino acid changes 
across all tiles}
\usage{
drawCoverageMap(
  params,
  info,
  aaMarginal,
  wmThreshold,
  tilesPerRegion,
  tileDepths,
  depthTable,
  tileCensi,
  tileLambdas
)
}
\arguments{
\item{params}{parameter sheet object}

\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{aaMarginal}{marginal frequency table at amino acid-level}

\item{wmThreshold}{'well-measuredness' threshold}

\item{tilesPerRegion}{a list of vectors containing the tile IDs for each region 'ri'}

\item{tileDepths}{table with adjusted mean effective sequencing depth for each tile}

\item{depthTable}{raw depth table}

\item{tileCensi}{table containing population breakdown of variant types within each tile}

\item{tileLambdas}{average number of mutations in each tile}
}
\description{
Output heatmap of marginal frequencies for all possible amino acid changes 
across all tiles
}
