% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{drawDistributions}
\alias{drawDistributions}
\title{Delegation function to draw score distribution plots with a given filter setting}
\usage{
drawDistributions(scores, sdCutoff)
}
\arguments{
\item{scores}{the score table}

\item{sdCutoff}{the stdev cutoff to apply}
}
\description{
Delegation function to draw score distribution plots with a given filter setting
}
