% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{drawNucleotideBias}
\alias{drawNucleotideBias}
\title{Draw Nucleotide Bias Analysis plots}
\usage{
drawNucleotideBias(info, params, simplifiedMarginal, tile)
}
\arguments{
\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{params}{parameter sheet object}

\item{simplifiedMarginal}{simplified marginal frequency table}

\item{tile}{the tile number being analyzed}
}
\description{
Draw Nucleotide Bias Analysis plots
}
