% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{drawPositionalDepth}
\alias{drawPositionalDepth}
\title{Output effective depth plot with effective sequencing depth across the tiles 
for each condition and replicate}
\usage{
drawPositionalDepth(covTable, outDir, pdftag)
}
\arguments{
\item{covTable}{dataframe with position-specific effective depth at each 
nucleotide position in each sample}

\item{outDir}{output directory}

\item{pdftag}{PDF tag string}
}
\description{
Output effective depth plot with effective sequencing depth across the tiles 
for each condition and replicate
}
