% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropoutAnalysis.R
\name{dropoutAnalysis}
\alias{dropoutAnalysis}
\title{run variant dropout troubleshooter}
\usage{
dropoutAnalysis(
  dataDir,
  countDir = NA,
  outDir = NA,
  paramFile = paste0(dataDir, "parameters.json"),
  srOverride = FALSE,
  tile = NA
)
}
\arguments{
\item{dataDir}{working data directory}

\item{countDir}{input directory for counts, defaults to subdirectory with latest timestamp ending in _mut_count.}

\item{outDir}{output directory, defaults to name of input directory with _QC tag attached.}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}

\item{srOverride}{single replicate override (use with caution)}
}
\value{
NULL. Results are written to file.
}
\description{
run variant dropout troubleshooter
}
