% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{enactScale}
\alias{enactScale}
\title{Calculate scores by scaling bce values to syn/nonsense medians}
\usage{
enactScale(msc, aac, sdThreshold, overrides = c(syn = NA, non = NA))
}
\arguments{
\item{msc}{data.frame containing the enrichment ratios (phi) and log(phi)}

\item{aac}{the vector of corresponding amino acid changes}

\item{sdThreshold}{stdev threshold for finding the syn/stop means}
}
\value{
data.frame with variant type, score and stdev
}
\description{
Calculate scores by scaling bce values to syn/nonsense medians
}
