% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{errorProfile}
\alias{errorProfile}
\title{Draw error profile for the dataset}
\usage{
errorProfile(scores, sCond, tp, outDir, params)
}
\arguments{
\item{scores}{the score table}

\item{sCond}{the condition for which to draw the plot}

\item{tp}{the time point}

\item{outDir}{the output directory}
}
\description{
Draw error profile for the dataset
}
