% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{expectedError}
\alias{expectedError}
\title{Infer a function that models expected error relative to data point magnitutes}
\usage{
expectedError(values, errors, mirror = FALSE, wtX = 0, intervalWidth = 0.1)
}
\arguments{
\item{values}{underlying datapoints}

\item{errors}{the error associated with each datapoint}
}
\value{
a list containing (i) a model function that maps data point magnitude to 
  expected error; (ii) the model parameters; and (iii) the a running median error across 
  the provided value range.
}
\description{
Infer a function that models expected error relative to data point magnitutes
}
