% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{exportExperiment}
\alias{exportExperiment}
\title{Export simulated experiment}
\usage{
exportExperiment(reads, params, outdir)
}
\arguments{
\item{reads}{list of sample read counts with names corresponding to samples}

\item{params}{the parameter object}

\item{outdir}{the target output folder to export to}
}
\value{
nothing
}
\description{
Export simulated experiment
}
