% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{exportRawAssay}
\alias{exportRawAssay}
\title{Exports raw assay data}
\usage{
exportRawAssay(assayReps, outdir)
}
\arguments{
\item{assayReps}{the assay replicates object}

\item{outdir}{the output directory}
}
\value{
nothing
}
\description{
Export the raw number of cells in each assay replicate
(before tileseq and sequencencing)
}
