% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{exportSeqSample}
\alias{exportSeqSample}
\title{Export count data to file}
\usage{
exportSeqSample(reads, name, outdir, params, hgvsb = new.hgvs.builder.c())
}
\arguments{
\item{reads}{the read information}

\item{name}{the sample name (matching the sample sheet)}

\item{outdir}{the output directory}

\item{params}{the parameter sheet}

\item{hgvsb}{a (coding sequence) HGVS builder object}
}
\value{

}
\description{
Export count data to file
}
