% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{filterProgression}
\alias{filterProgression}
\title{Draw replicate correlation plots}
\usage{
filterProgression(scores, sCond, tp, params, outDir)
}
\arguments{
\item{scores}{the score table}

\item{params}{the global parameter object}
}
\description{
Draw replicate correlation plots
}
