% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{filterVariants}
\alias{filterVariants}
\title{Filter variants that have greater depth drop than max allowed depth drop or 
do not have minimum required depth}
\usage{
filterVariants(
  nsCond,
  params,
  xDepthDrops,
  xCounts,
  xSplitChanges,
  xTiles,
  mc.cores,
  allVariants = TRUE
)
}
\arguments{
\item{nsCond}{the nonselect condition}

\item{params}{parameter sheet object}

\item{xDepthDrops}{table listing the relative drop for variants in xCounts table}

\item{xCounts}{marginal or all counts table}

\item{xSplitChanges}{table with split codon changes for variants in xCounts}

\item{xTiles}{tile assignments for each variant in xCounts table}

\item{mc.cores}{number of processes to run in parallel for multi-core processing}

\item{allVariants}{flag, whether the xCounts table contains all variants}
}
\description{
Filter variants that have greater depth drop than max allowed depth drop or 
do not have minimum required depth
}
