% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{findQuads}
\alias{findQuads}
\title{Helper function to find the quad of conditions and time points that belong 
with the given selection condition}
\usage{
findQuads(params, sCond, tp)
}
\arguments{
\item{params}{the parameter sheet object}

\item{sCond}{the current selection condition}

\item{tp}{the current time point of the selection condition}
}
\value{
a list containing the condition quad ('condQuad'), 
replicate matrix ('repMatrix') and timepoint quad ('tpQuad')
}
\description{
Helper function to find the quad of conditions and time points that belong 
with the given selection condition
}
