% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{findRuns}
\alias{findRuns}
\title{Finds runs of TRUE values in a boolean vector}
\usage{
findRuns(bools)
}
\arguments{
\item{bools}{a logical vector}
}
\value{
an integer matrix with columns 'start' and 'end' indicating runs
}
\description{
Finds runs of TRUE values in a boolean vector
}
