% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{fit.cv.model}
\alias{fit.cv.model}
\title{Create a model of the error in for a subset of marginal frequencies}
\usage{
fit.cv.model(subscores, cond)
}
\arguments{
\item{subcores}{a subset of the count table (for a given tile)}
}
\value{
a list containing the model function (mapping counts to expected CV), and the model parameters
}
\description{
Create a model of the error in for a subset of marginal frequencies
}
