% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{floor0}
\alias{floor0}
\title{Helper function to bottom-out a vector of numbers, so that no element is smaller than 'bottom'}
\usage{
floor0(xs, bottom = 0)
}
\arguments{
\item{xs}{the input numerical vector}

\item{bottom}{the minimum to which the vector will be forced.}
}
\description{
Helper function to bottom-out a vector of numbers, so that no element is smaller than 'bottom'
}
