% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{flooring}
\alias{flooring}
\title{Apply flooring and SD adjustment to negative scores}
\usage{
flooring(msc, params)
}
\arguments{
\item{msc}{the score matrix}

\item{params}{the global parameter object}
}
\value{
a matrix containing the floored scores and SDs
}
\description{
This assumes that no cell can be "deader than dead", and that negative scores are just due to 
measurement error. From this follows that the more negative the score, the more certain we 
can be that the cell is dead. We can therefore adjust the SD to reflect this, by setting 
the score to zero and shifting the SD such that the implied probability of being above 0 
remains the same.
}
