% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorScore.R
\name{fluorScore}
\alias{fluorScore}
\title{score fluorescence data}
\usage{
fluorScore(
  dataDir,
  conditionOrder,
  paramFile = paste0(dataDir, "parameters.json"),
  srOverride = FALSE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{conditionOrder}{vector containting the order in which the conditions represent fluorescence bins}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}
}
\value{
NULL. Results are written to file.
}
\description{
score fluorescence data
}
