% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{getNonselectFor}
\alias{getNonselectFor}
\title{Convenience function get the matching non-selective condition for the given input condition}
\usage{
getNonselectFor(cond, params)
}
\arguments{
\item{cond}{the input condition}

\item{param}{the parameter object}
}
\value{
the list of matching non-selective conditions
}
\description{
Convenience function get the matching non-selective condition for the given input condition
}
