% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{getNonselects}
\alias{getNonselects}
\title{Convenience function get all non-selective conditions}
\usage{
getNonselects(params)
}
\arguments{
\item{param}{the parameter object}
}
\value{
the list of non-selective conditions
}
\description{
Convenience function get all non-selective conditions
}
