% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{getNormOverrides}
\alias{getNormOverrides}
\title{Helper function to extract normalization override values from the parameter sheet
if they exist}
\usage{
getNormOverrides(params, sCond, tp, region)
}
\arguments{
\item{params}{the parameter sheet}

\item{sCond}{the current selective condition ID}

\item{tp}{the time point ID}

\item{region}{the region ID}
}
\value{
a vector with syn and stop overrides, or \code{NA} if they weren't defined.
}
\description{
Helper function to extract normalization override values from the parameter sheet
if they exist
}
