% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{getNsConditions}
\alias{getNsConditions}
\title{Get all nonselect conditions from parameter sheet object}
\usage{
getNsConditions(params, allCondOverride)
}
\arguments{
\item{params}{parameter sheet object}

\item{allCondOverride}{override flag to run on ALL conditions instead of just
nonselect}
}
\value{
list of nonselect conditions
}
\description{
Get all nonselect conditions from parameter sheet object
}
