% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{getSelects}
\alias{getSelects}
\title{Convenience function get all selective conditions}
\usage{
getSelects(params)
}
\arguments{
\item{param}{the parameter object}
}
\value{
the list of selective conditions
}
\description{
Convenience function get all selective conditions
}
