% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{getTimepointsFor}
\alias{getTimepointsFor}
\title{Convenience function get the available time points for the given input condition}
\usage{
getTimepointsFor(cond, params)
}
\arguments{
\item{cond}{the input condition}

\item{param}{the parameter object}
}
\value{
the list of available time points
}
\description{
Convenience function get the available time points for the given input condition
}
