% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{getWTControlFor}
\alias{getWTControlFor}
\title{Convenience function get the matching WT control condition for the given input condition}
\usage{
getWTControlFor(cond, params)
}
\arguments{
\item{param}{the parameter object}
}
\value{
the list of matching WT control conditions
}
\description{
Convenience function get the matching WT control condition for the given input condition
}
