% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{getWTReps}
\alias{getWTReps}
\title{Extract appropriate WT replicates for timepoint given in 'info'}
\usage{
getWTReps(info, params, wtCond)
}
\arguments{
\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{params}{parameter sheet object}

\item{wtCond}{}
}
\value{
vector of strings with corresponding WT replicates
}
\description{
Extract appropriate WT replicates for timepoint given in 'info'
}
