% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{inferTiles}
\alias{inferTiles}
\title{Determine which tiles the positions from split codon changes table 'cct' 
belong to}
\usage{
inferTiles(params, cct)
}
\arguments{
\item{params}{parameter sheet object}

\item{cct}{table with split codon changes}
}
\value{
vector of corresponding tile numbers for each codon change
}
\description{
Determine which tiles the positions from split codon changes table 'cct' 
belong to
}
