% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{join.datapoints}
\alias{join.datapoints}
\title{join multiple datapoints weighted by stdev}
\usage{
join.datapoints(ms, sds, dfs)
}
\arguments{
\item{ms}{data means}

\item{sds}{data stdevs}

\item{degrees}{of freedom for each data point}
}
\value{
a vector containing the joint mean and joint stdev
}
\description{
join multiple datapoints weighted by stdev
}
