% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRawInputs.R
\name{joinRawInputs}
\alias{joinRawInputs}
\title{join raw input datasets}
\usage{
joinRawInputs(countfiles, outfile = NULL, logger = NULL)
}
\arguments{
\item{countfiles}{A vector of filenames for the input count files}

\item{outfile}{The path of the output file. Defaults to NULL, in which case the
output is written to a file in the same directory as the firsts input file, named
similarly to said input file with an added "_merged" tag.}

\item{logger}{A yogilogger object, may be omitted, in which case messages are written
directly to stdout.}
}
\value{
the merged data as a \code{data.frame}.
}
\description{
Joins multiple input datasets into a single one, treating each as additional replicates
}
\examples{

}
