% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{latestSubDir}
\alias{latestSubDir}
\title{Convenience function to find a subdirectory matching the given pattern.}
\usage{
latestSubDir(parentDir, pattern = "_mut_call$|mut_count$")
}
\arguments{
\item{parentDir}{the parent folder in which to look for matching subfolders}

\item{pattern}{a regex pattern used to identify the correct subfolder}
}
\value{
a vector containing the path to the subfolder, its timestamp and any possible label.
}
\description{
The subfolders should follow the scheme [<label>_]<timestamp>_<outputType>/
}
