% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{libraryQCLegacy}
\alias{libraryQCLegacy}
\title{QC plots for library coverage}
\usage{
libraryQCLegacy(
  dataDir,
  outdir,
  mut2funcFile = paste0(dataDir, "mut2func_info.csv"),
  logger = NULL,
  mc.cores = 8
)
}
\arguments{
\item{dataDir}{path to the existing data main directory}

\item{outdir}{path to the output directory}

\item{mut2funcFile}{path to the mut2func file. Defaults to <dataDir>/mut2func_info.csv}

\item{logger}{optional yogilogger object. Defaults to NULL, in which case messages will go to stdout.}

\item{mc.cores}{number of CPU cores to use in parallel}
}
\value{
NULL. Results are written to file.
}
\description{
QC plots for library coverage
}
