% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{log.sd}
\alias{log.sd}
\title{Error propagation formula for the logarithm of a Random Variable}
\usage{
\method{log}{sd}(m1, sd1, base = 10)
}
\arguments{
\item{m1}{the mean of the random variable}

\item{sd1}{the standard deviation of the RV}

\item{base}{the logarithm base (default 10)}
}
\description{
Error propagation formula for the logarithm of a Random Variable
}
