% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{logPhiBias}
\alias{logPhiBias}
\title{Plot the logPhi bias vs the nonselect marginal frequency}
\usage{
logPhiBias(scores, params, sCond, tp, outDir)
}
\arguments{
\item{scores}{the score table}

\item{params}{the parameter sheet object}

\item{sCond}{the active selection condition}

\item{tp}{the active time point}

\item{outDir}{the output directory}
}
\value{
nothing. writes a pdf file to the output dir
}
\description{
Plot the logPhi bias vs the nonselect marginal frequency
}
