% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logVersion}
\alias{logVersion}
\title{Convenience function for recording the package version into the logfile.}
\usage{
logVersion()
}
\description{
Convenience function for recording the package version into the logfile.
}
