% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logWarn}
\alias{logWarn}
\title{convenience logging function at WARN level}
\usage{
logWarn(...)
}
\arguments{
\item{...}{any objects or messages to be logged}
}
\value{
nothing
}
\description{
Checks whether a logger has been registered globally and if so, uses it.
Otherwise writes the message to stdout.
}
