% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavedbCountExport.R
\name{mavedbCountExport}
\alias{mavedbCountExport}
\title{Convert count files from legacy pipeline to MaveDB format}
\usage{
mavedbCountExport(countfile, outdir, logger = NULL)
}
\arguments{
\item{countfile}{the path to the "rawData.txt" file produced by the legacy pipeline.}

\item{outdir}{path to desired output directory}
}
\value{
nothing. output is written to various files in the output directory
}
\description{
Convert count files from legacy pipeline to MaveDB format
}
