% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mean.sd.count}
\alias{mean.sd.count}
\title{Function to form mean, stdev and average raw count for a given condition}
\usage{
\method{mean}{sd.count}(cond, regionalCounts, regionalDrops, tp, params)
}
\arguments{
\item{cond}{the names of the condition (e.g. select, nonselect)}

\item{regionalCounts}{the region-specific subset of the marginal counts dataframe}

\item{tp}{the time point ID}

\item{params}{the global parameters object}
}
\description{
Function to form mean, stdev and average raw count for a given condition
}
