% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{mut_breakdown}
\alias{mut_breakdown}
\title{Return the sums of marginal frequencies for different types of mutations in 
each tile}
\usage{
mut_breakdown(tile, simplifiedMarginal, freq, fromaa, toaa, to)
}
\arguments{
\item{tile}{tile number being analyzed}

\item{simplifiedMarginal}{simplified marginal frequency table}

\item{freq}{vector of frequencies for each mutation in simplifiedMarginal}

\item{fromaa}{starting amino acid}

\item{toaa}{ending amino acid}

\item{to}{length of codon change ****}
}
\value{
list with the sums of marginal frequencies for different mutations types
}
\description{
Return the sums of marginal frequencies for different types of mutations in 
each tile
}
