% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{normalizeScores}
\alias{normalizeScores}
\title{Calculate scores by normalizing logPhi values to syn/nonsense medians}
\usage{
normalizeScores(msc, aac, sdThreshold)
}
\arguments{
\item{msc}{data.frame containing the enrichment ratios (phi) and log(phi)}

\item{aac}{the vector of corresponding amino acid changes}

\item{sdThreshold}{the maximum stdev of logPhi to filter for before calculating medians}
}
\value{
data.frame with variant type, score and stdev
}
\description{
Calculate scores by normalizing logPhi values to syn/nonsense medians
}
