% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{nucleotideBiasAnalysis}
\alias{nucleotideBiasAnalysis}
\title{Helper function to perform nucleotide bias analysis and draw the corresponding plot}
\usage{
nucleotideBiasAnalysis(simplifiedMarginal, tile, draw = TRUE)
}
\arguments{
\item{simplifiedMarginal}{simplified marginal frequency table}

\item{tile}{the tile number being analyzed}

\item{draw}{flag, whether to draw nucleotide bias analysis plot}
}
\value{
rates at which each nucelotide is substituted for another across the tile
}
\description{
Helper function to perform nucleotide bias analysis and draw the corresponding plot
}
