% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateHGVS.R
\name{parseCountFile}
\alias{parseCountFile}
\title{parse a count table}
\usage{
parseCountFile(filename)
}
\arguments{
\item{filename}{input file}
}
\value{
data.frame with attributes containing header values
}
\description{
parse a count table
}
