% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateHGVS.R
\name{parseCoverageFile}
\alias{parseCoverageFile}
\title{parse coverage data file}
\usage{
parseCoverageFile(filename, params, tile)
}
\arguments{
\item{filename}{the input csv file}

\item{params}{the parameter sheet}

\item{tile}{the tile number associated with this file}
}
\value{
a vector detailing the number of rejected reads for each position
}
\description{
parse coverage data file
}
