% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{parseParameters}
\alias{parseParameters}
\title{parse JSON parameter file}
\usage{
parseParameters(filename)
}
\arguments{
\item{filename}{the input CSV file}
}
\value{
the parameter object, as a list of lists
}
\description{
parse JSON parameter file
}
