% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{pdftagger}
\alias{pdftagger}
\title{Helper object for adding tags to pdf page margins}
\usage{
pdftagger(tag, cpp = 1)
}
\arguments{
\item{tag}{str, description tag to be added}

\item{cpp}{number of cycles per page}
}
\value{
list that contains functions for adding tags to page margins
}
\description{
Helper object for adding tags to pdf page margins
}
