% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotCensus}
\alias{plotCensus}
\title{Helper function to plot a census dataset}
\usage{
plotCensus(census, lambda, d = NULL, e = NULL, main = "")
}
\arguments{
\item{census}{table of population breakdown of variants within tile/region}

\item{d}{raw depth}

\item{e}{effective depth}

\item{main}{str, title of plot}

\item{lamda}{average number of mutations in tile/region}
}
\description{
Helper function to plot a census dataset
}
