% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotCoverageCurve}
\alias{plotCoverageCurve}
\title{Output coverage curves showing the progression of how many variants are observed at 
increasing marginal frequency thresholds}
\usage{
plotCoverageCurve(coverageCurves, thresholds, ri, wmThreshold, tagger)
}
\arguments{
\item{coverageCurves}{dataframe with cumulative distribution of variants observed 
at different marginal frequency thresholds}

\item{thresholds}{colour transition thresholds}

\item{ri}{region number}

\item{wmThreshold}{well-measuredness threshold}

\item{tagger}{list that contains functions for adding tags to page margins}
}
\description{
Output coverage curves showing the progression of how many variants are observed at 
increasing marginal frequency thresholds
}
