% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotDepthDrops}
\alias{plotDepthDrops}
\title{Output depth drop distribution plot showing distribution of relative 
drops in effective sequencing depth}
\usage{
plotDepthDrops(params, outDir, pdftag, allDepthDrops)
}
\arguments{
\item{params}{parameter sheet object}

\item{allDepthDrops}{table listing the relative drop for all variants in 
each condition-time-replicate combination}

\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}
}
\description{
Output depth drop distribution plot showing distribution of relative 
drops in effective sequencing depth
}
