% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotFrameshiftHotspot}
\alias{plotFrameshiftHotspot}
\title{Output frameshift hotspot plot with distribution of frameshift variants across 
template sequence}
\usage{
plotFrameshiftHotspot(
  info,
  params,
  fsIdx,
  marginalSplitChanges,
  nsMarginalMeans,
  marginalCounts
)
}
\arguments{
\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{params}{parameter sheet object}

\item{fsIdx}{frameshift indices from marginal counts table}

\item{marginalSplitChanges}{table with split codon changes from marginal counts table}

\item{nsMarginalMeans}{table with mean marginal frequency across nonselect technical replicates}

\item{marginalCounts}{table of marginal counts}
}
\description{
Output frameshift hotspot plot with distribution of frameshift variants across 
template sequence
}
