% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotJackpotDiagnosis}
\alias{plotJackpotDiagnosis}
\title{Output jackpot plot with marginal frequencies of all variants in decreasing order}
\usage{
plotJackpotDiagnosis(info, nsMarginalMeans, wmThreshold, marginalCounts)
}
\arguments{
\item{nsMarginalMeans}{table with mean marginal frequency across nonselect technical replicates}

\item{wmThreshold}{well-measuredness threshold}

\item{marginalCounts}{table of marginal counts}
}
\description{
Output jackpot plot with marginal frequencies of all variants in decreasing order
}
