% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotMutbreakdown}
\alias{plotMutbreakdown}
\title{Output variant-type breakdown for each tile}
\usage{
plotMutbreakdown(info, mutbreakdown)
}
\arguments{
\item{info}{list containing condition details (nscond), timepoint (tp), 
pdftag and outdir}

\item{mutbreakdown}{table containing the sums of marginal frequencies 
for different mutation types}
}
\description{
Output variant-type breakdown for each tile
}
