% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotReplicateCorrelation}
\alias{plotReplicateCorrelation}
\title{Output replicate correlation plot between technical replicates from nonselect condition}
\usage{
plotReplicateCorrelation(params, nsReps, info, marginalCounts, marginalTiles)
}
\arguments{
\item{params}{parameter sheet object}

\item{nsReps}{number of nonselect replicates}

\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{marginalCounts}{marginal counts table}

\item{marginalTiles}{table with tile assignments for each entry in marginal 
counts table}
}
\description{
Output replicate correlation plot between technical replicates from nonselect condition
}
