% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotSeqErrorRates}
\alias{plotSeqErrorRates}
\title{Output basecall quality plot with observed error rate at given basecall PHRED 
quality score}
\usage{
plotSeqErrorRates(inDir, outDir, pdftag)
}
\arguments{
\item{inDir}{input directory}

\item{outDir}{output directory}

\item{pdftag}{PDF tag string}
}
\description{
Output basecall quality plot with observed error rate at given basecall PHRED 
quality score
}
