% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{plotWTCtrl}
\alias{plotWTCtrl}
\title{Draw WT frequency plot comparing distribution of marginal frequencies in 
nonselect condition against WT condition}
\usage{
plotWTCtrl(
  info,
  params,
  marginalTiles,
  wtMarginalMeans,
  nsMarginalMeans,
  pseudoCount
)
}
\arguments{
\item{info}{list containing details on condition (nscond), timepoint (tp), 
pdftag and outdir}

\item{params}{parameter sheet object}

\item{marginalTiles}{table with tile assignments for each entry in marginal 
counts table}

\item{wtMarginalMeans}{table with mean marginal frequency across technical 
replicates in WT condition}

\item{nsMarginalMeans}{table with mean marginal frequency across technical 
replicates in nonselect condition}

\item{pseudoCount}{a small number added to each frequency to allow for log-scale visualization}
}
\description{
Draw WT frequency plot comparing distribution of marginal frequencies in 
nonselect condition against WT condition
}
