% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ratio.sd}
\alias{ratio.sd}
\title{Error propagation formula for ratios of Random Variables}
\usage{
ratio.sd(m1, m2, sd1, sd2, cov12 = 0)
}
\arguments{
\item{m1}{the mean of the numerator RV}

\item{m2}{the mean of the denominator RV}

\item{sd1}{the stdev of the numerator RV}

\item{sd2}{the stdev of the denominator RV}

\item{cov12}{the covariance of the two variables}
}
\description{
Error propagation formula for ratios of Random Variables
}
