% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{reachableChanges}
\alias{reachableChanges}
\title{Convenience function to calculate the list of SNV-reachable AA changes}
\usage{
reachableChanges(params)
}
\arguments{
\item{param}{the parameter object}
}
\value{
a data.frame containing wt and mutant codons as well as wt and mutant AAs for all possible SNVs
}
\description{
Given a parameter object containing a coding sequence, this function will calculate a table
of all SNV-reachable amino acid changes, detailing which codon changes correspond to each.
}
