% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{regionCensus}
\alias{regionCensus}
\title{Perform census on overall region to determine what mutations are present}
\usage{
regionCensus(ri, tilesPerRegion, tileCensi, tileLambdas, maxChanges, tagger)
}
\arguments{
\item{ri}{region index; region being analyzed}

\item{tilesPerRegion}{a list of vectors containing the tile IDs for each region 'ri'}

\item{tileCensi}{table containing population breakdown of variant types within each tile}

\item{tileLambdas}{average number of mutations in each tile}

\item{maxChanges}{maximum number of mutations allowed on x-axis of plot}

\item{tagger}{list that contains functions for adding tags to page margins}
}
\value{
table containing population breakdown of variant types within region
}
\description{
Perform census on overall region to determine what mutations are present
}
