% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{registerLogger}
\alias{registerLogger}
\title{Globally registers a logger object}
\usage{
registerLogger(logger)
}
\arguments{
\item{logger}{a yogilogger object}
}
\value{
nothing
}
\description{
Globally registers a logger object
}
