% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{regularizationQC}
\alias{regularizationQC}
\title{Draw error regulariztion model QC plots}
\usage{
regularizationQC(scores, modelParams, params, sCond, tp, outDir)
}
\arguments{
\item{scores}{the score table}

\item{params}{the global parameter object}

\item{sCond}{the condition for which to draw the plot}

\item{tp}{the time point}

\item{outDir}{the output directory}
}
\description{
Draw error regulariztion model QC plots
}
