% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{regularizeRaw}
\alias{regularizeRaw}
\title{Function to perform Baldi & Long's error regularization}
\usage{
regularizeRaw(
  msc,
  condNames,
  tiles,
  n,
  pseudo.n,
  modelFunctions,
  pessimistic = TRUE
)
}
\arguments{
\item{msc}{dataframe containing the output of the 'mean.sd.count()' function}

\item{condNames}{the names of the different count conditions (select, nonselect etc)}

\item{n}{the number of replicates present for each datapoint in each condition}

\item{pseudo.n}{the number of pseudo-replicates that detemine the weight of the prior
in the regularization}
}
\value{
a data.frame containing for each condition the possion prior, bayesian regularized, and
   pessimistic (maximum) standard deviation.
}
\description{
Function to perform Baldi & Long's error regularization
}
