% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{replicateCorrelation}
\alias{replicateCorrelation}
\title{Draw replicate correlation plots}
\usage{
replicateCorrelation(scores, marginalCounts, params, sCond, tp, outDir)
}
\arguments{
\item{scores}{the score table}

\item{marginalCounts}{the marginal counts table}

\item{params}{the global parameter object}

\item{sCond}{the condition for which to draw the plot}

\item{tp}{the time point}

\item{outDir}{the output directory}
}
\description{
Draw replicate correlation plots
}
