% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{runModelFits}
\alias{runModelFits}
\title{Run model fitting on all tiles in all given conditions}
\usage{
runModelFits(msc, condNames, tiles, mc.cores)
}
\arguments{
\item{msc}{dataframe containing the output of the 'mean.sd.count()' function.}

\item{condNames}{the condition names}

\item{tiles}{vector of tile assignments for each row in msc}

\item{mc.cores}{the number of CPU cores to use in parallel}
}
\description{
Run model fitting on all tiles in all given conditions
}
