% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{scoreDistributions}
\alias{scoreDistributions}
\title{Draw score distribution plots}
\usage{
scoreDistributions(scores, sCond, tp, outDir, sdCutoff)
}
\arguments{
\item{scores}{the score table}

\item{sCond}{the condition for which to draw the plot}

\item{tp}{the time point}

\item{outDir}{the output directory}
}
\description{
Draw score distribution plots
}
