% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateExperiment}
\alias{simulateExperiment}
\title{Title}
\usage{
simulateExperiment(
  workdir,
  paramFile = paste0(workdir, "parameters.json"),
  poolCV = 0.2,
  poolLambda = 1,
  fitBias = 0,
  fitSpread = 4,
  nPlasmids = 5e+05,
  nClones = 5e+05,
  assaySD = 0.01,
  assayTime = 3,
  seqDepth = 5e+06,
  varcallErr = 3e-06
)
}
\arguments{
\item{workdir}{current working directory}

\item{paramFile}{parameter file}

\item{poolCV}{coefficient of variation in original mutant pool}

\item{poolLambda}{number of mutations per clone}

\item{fitBias}{fitness bias (towards neutral or deleterious)}

\item{fitSpread}{fitness spread (how strongly it tends to extremes)}

\item{nPlasmids}{size of initial plasmid pool}

\item{nClones}{number of clones to be used in each replicate pool}

\item{assaySD}{the amount of noise in the assay}

\item{assayTime}{the growth time in the assay}

\item{seqDepth}{the sequencing depth for tileseq}

\item{varcallErr}{the per-base variant calling error}
}
\value{

}
\description{
Title
}
