% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateFitness}
\alias{simulateFitness}
\title{Simulate Clone fitness}
\usage{
simulateFitness(params, pool, fitBias = 0, fitSpread = 4)
}
\arguments{
\item{params}{The parameter sheet}

\item{pool}{the output of simulatePool()}

\item{lofFreq}{the frequency of loss-of-function effects to simulate}

\item{hypoFreq}{the frequency of hypomorphic effects to simulate}
}
\description{
This function generates assay fitness values for clones and their variants.
It outputs a list with two elements: (i) a table listing the fitness effects of
All possible nucleotide changes, and (ii) a function that calculates the true fitness
of any given clone given its variant combination. Currently, this does not simulate
genetic interactions, but simply uses multiplicativity of fitness. It also currently 
does not simulate hypercomplementation effects.
}
