% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulatePool}
\alias{simulatePool}
\title{Simulate Pool}
\usage{
simulatePool(params, poolCV = 0.2, poolLambda = 1)
}
\arguments{
\item{params}{a parameter sheet to follow (in terms of CDS sequence etc.)}

\item{poolCV}{the coefficient of variation for the distribution of marginal frequencies}

\item{poolLambda}{the mean of the poisson distibution used to sample the number of 
codon-changes per clone}
}
\description{
This simulates a pool of variant-bearing clones. The output is a list containing
two elements: (i) a table of all underlying codon changes and their true marginal
frequencies, and (ii) a function that samples (multi-variant-bearing) clone populations
on demand
}
