% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateTileSeq}
\alias{simulateTileSeq}
\title{simulate TileSeq}
\usage{
simulateTileSeq(assayReps, params, depth = 1e+05, varcallErr = 3e-06)
}
\arguments{
\item{assayReps}{a list of clone pool replicates. These can be pre- or post-selection
or even all WT}

\item{params}{the parameter sheet}

\item{depth}{the sequencing depth to simulate}

\item{varcallErr}{the variant calling error to simulate}
}
\description{
simulate TileSeq
}
