% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{slide}
\alias{slide}
\title{Underlying function for error profile model}
\usage{
slide(xs, x0, y0, x1, y1, s0, s1)
}
\arguments{
\item{xs}{the input x-values}

\item{x0}{the starting x of the slide}

\item{y0}{the starting y value of the slide (the constant level on the left)}

\item{x1}{the end x of the slide}

\item{y1}{the end y of the slide (i.e. the constant level on the right)}

\item{s0}{the amount of smoothing on the left}

\item{s1}{the amount of smoothing on the right}
}
\value{
the y values corresponding to the input x
}
\description{
Creates a function that slides smoothly from one constant level to another
}
