% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{splitCodonChanges}
\alias{splitCodonChanges}
\title{Extract affected positions and codon details from vector 'css'}
\usage{
splitCodonChanges(ccs)
}
\arguments{
\item{ccs}{vector containing codon changes}
}
\value{
table containing start amino acid, position, and end amino acid
}
\description{
Extract affected positions and codon details from vector 'css'
}
