% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{synNonDelta}
\alias{synNonDelta}
\title{Draws a plot showing running means of synonymous and nonsense variant scores
across the length of the protein, as well as difference track between the two.}
\usage{
synNonDelta(scores, sCond, tp, params, outDir)
}
\arguments{
\item{scores}{data.frame of the scores}

\item{sCond}{the currently active selective condition}

\item{tp}{the currently active time point}

\item{params}{the parameter sheet object}

\item{outDir}{the output directory}
}
\value{
nothing, writes plot to output directory
}
\description{
Draws a plot showing running means of synonymous and nonsense variant scores
across the length of the protein, as well as difference track between the two.
}
