% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{tileCensus}
\alias{tileCensus}
\title{Perform census on tile to determine what mutations are present}
\usage{
tileCensus(
  tile,
  allTiles,
  maxChanges,
  nsAllMeans,
  isFrameshift,
  isInFrameIndel,
  numChanges
)
}
\arguments{
\item{tile}{the tile number being analyzed}

\item{allTiles}{table containing all tiles}

\item{maxChanges}{maximum number of mutations allowed on x-axis of plot}

\item{nsAllMeans}{table with mean frequency across nonselect technical replicates}

\item{isFrameshift}{vector containing flags of whether a variant is a frameshift}

\item{isInFrameIndel}{vector containing flags of whether a variant is an indel}

\item{numChanges}{vector containing number of nucleotide changes in a given mutational context}
}
\value{
table containing population breakdown of variant types within tile
}
\description{
Perform census on tile to determine what mutations are present
}
