% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{tilesInRegions}
\alias{tilesInRegions}
\title{Convenience function to find the tiles that belong to each region.}
\usage{
tilesInRegions(params)
}
\arguments{
\item{param}{the parameter object}
}
\value{
a  list of vectors containing the tile IDs for each region ID
}
\description{
Given the parameter object returns a list of vectors containing the tile IDs for each region ID
}
