% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateHGVS.R
\name{translateHGVS}
\alias{translateHGVS}
\title{Translate a single HGVS string from nucleotide to amino acid level}
\usage{
translateHGVS(hgvs, params, builder = new.hgvs.builder.p(aacode = 3),
  strictMode = TRUE)
}
\arguments{
\item{hgvs}{a single cds-level HGVS string. May contain multiple mutations in-cis.}

\item{params}{a parameter object with CDS definitions}

\item{builder}{an optional protein HGVS builder object, if none is provided, a new one is created.}

\item{strictMode}{whether HGVS will be validated against the reference sequence before translation}
}
\value{
translated HGVS string
}
\description{
Translate a single HGVS string from nucleotide to amino acid level
}
