% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildJointTable.R
\name{buildJointTable}
\alias{buildJointTable}
\title{run library quality control (QC)}
\usage{
buildJointTable(
  dataDir,
  inDir = NA,
  outDir = NA,
  paramFile = paste0(dataDir, "parameters.json"),
  mc.cores = 6,
  srOverride = FALSE,
  covOverride = FALSE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{inDir}{input directory, defaults to subdirectory with latest timestamp ending in _mut_count.}

\item{outDir}{output directory, defaults to input directory.}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}

\item{mc.cores}{the number of CPU cores to use in parallel}

\item{srOverride}{the single-replicate override flag}

\item{covOverride}{override the requirement for coverage files. For backwards-
compatibility with older versions prior to v0.7}
}
\value{
NULL. Results are written to file.
}
\description{
run library quality control (QC)
}
