% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryQC.R
\name{calcDepthDrops}
\alias{calcDepthDrops}
\title{Calculate relative drops in effective sequencing depth for each variant}
\usage{
calcDepthDrops(xCounts, xTiles, depthTable, mc.cores = 6)
}
\arguments{
\item{xCounts}{marginal or all counts table}

\item{xTiles}{tile assignments for each variant in the table}

\item{depthTable}{the raw depth table}
}
\value{
a table listing the relative drop in for each variant in each
condition-time-replicate combination.
}
\description{
Calculate relative drops in effective sequencing depth for each variant
}
