% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{collapseByAA}
\alias{collapseByAA}
\title{Collapse score table by amino acid changes}
\usage{
collapseByAA(
  scoreTable,
  params,
  sCond,
  scoreCol = "score",
  sdCol = "score.sd",
  srOverride = FALSE
)
}
\arguments{
\item{scoreTable}{the full score}

\item{params}{the parameter object}

\item{sCond}{the current selective condition}

\item{scoreCol}{the name of the column containing the scores}

\item{sdCol}{the name of the column containing the stdev}

\item{sdOverride}{the sdOverride flag}
}
\value{
a \code{data.frame} containing the collapsed score table
}
\description{
Collapse score table by amino acid changes
}
