% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{csvParam2Json}
\alias{csvParam2Json}
\title{convert CSV input parameter file to JSON format}
\usage{
csvParam2Json(infile, outfile = sub("[^/]+$", "parameters.json",
  filename), logger = NULL)
}
\arguments{
\item{infile}{the input CSV file}

\item{outfile}{the output JSON file. Defaults to parameters.json in the same directory}

\item{logger}{yogilog logger. Defaults to NULL and writes to stdout.}
}
\value{
NULL. Results are written to file.
}
\description{
convert CSV input parameter file to JSON format
}
