% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{drawDistributions}
\alias{drawDistributions}
\title{Delegation function to draw score distribution plots with a given filter setting}
\usage{
drawDistributions(aaScores, seCutoff = Inf, reg = NA, bcOverride = TRUE)
}
\arguments{
\item{aaScores}{the score table}

\item{seCutoff}{the stderr cutoff to apply}
}
\description{
Delegation function to draw score distribution plots with a given filter setting
}
