% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{expectedError}
\alias{expectedError}
\title{Infer a function that models expected error relative to data point magnitutes}
\usage{
expectedError(values, errors)
}
\arguments{
\item{values}{underlying datapoints}

\item{errors}{the error associated with each datapoint}
}
\value{
a function that maps data point magnitude to expected error
}
\description{
Infer a function that models expected error relative to data point magnitutes
}
