% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{join.datapoints}
\alias{join.datapoints}
\title{join multiple datapoints weighted by stdev}
\usage{
join.datapoints(ms, sds, dfs, ws = (1/sds)/sum(1/sds))
}
\arguments{
\item{ms}{data means}

\item{sds}{data stdevs}

\item{dfs}{degrees of freedom for each data point}

\item{ws}{datapoint weights. By default these get auto-calculated from sds}
}
\value{
a vector containing the joint mean and joint stdev
}
\description{
join multiple datapoints weighted by stdev
}
