% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{rawFilter}
\alias{rawFilter}
\title{Function to apply filtering based on raw counts / frequency}
\usage{
rawFilter(msc, countThreshold, wtq = 0.95)
}
\arguments{
\item{msc}{dataframe containing the output of the 'mean.sd.count()' function.}

\item{countThreshold}{the threshold of raw counts that must be met}
}
\value{
a vector listing for each row in the table which (if any) filters apply, otherwise NA.
}
\description{
Function to apply filtering based on raw counts / frequency
}
