% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoring}
\alias{scoring}
\title{run scoring function}
\usage{
scoring(
  dataDir,
  paramFile = paste0(dataDir, "parameters.json"),
  logger = NULL,
  mc.cores = 6,
  countThreshold = 10,
  pseudo.n = 8,
  sdThreshold = 0.3,
  srOverride = FALSE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}
}
\value{
NULL. Results are written to file.
}
\description{
run scoring function
}
