% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionQC.R
\name{selectionQC}
\alias{selectionQC}
\title{run library quality control (QC)}
\usage{
selectionQC(
  dataDir,
  paramFile = paste0(dataDir, "parameters.json"),
  logger = NULL,
  sdCutoff = 0.3,
  srOverride = TRUE
)
}
\arguments{
\item{dataDir}{working data directory}

\item{paramFile}{input parameter file. defaults to <dataDir>/parameters.json}
}
\value{
NULL. Results are written to file.
}
\description{
run library quality control (QC)
}
