% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvParam2Json.R
\name{validateParameters}
\alias{validateParameters}
\title{run validation checks on parameters object}
\usage{
validateParameters(params)
}
\arguments{
\item{params}{parameter object to validate}
}
\value{
TRUE if everything checks out, otherwise it throws errors
}
\description{
run validation checks on parameters object
}
